api_statistics/requirements.txt
asgiref==3.5.0
Django==3.2.12
pip==22.0.4
pytz==2022.1
setuptools==57.0.0
sqlparse==0.4.2
wheel==0.36.2
python-dotenv==0.20.0
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
django-filter==21.1